add blueprints
